requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-environ==0.9.0
idna==3.4
kombu==5.2.4
Pillow==9.3.0
pip==21.2.3
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pydantic==1.10.2
pytz==2022.7
redis==4.4.0
requests==2.28.1
setuptools==57.4.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
